PoC: Initial PoC is networking based
Given the sw_firewall PoC was initially committed over a month ago as a
part of ca9b736, with no additional work since then, we'll move to a
networking PoC for now. When the firewall PoC comes back to life, we can
add a directory for that one as well.

Signed-off-by: Kyle Mestery <mestery@mestery.com>