security: Change startup order
Sometimes, the vpn-client container fails to start with the following
message:

```
ERROR: for vpn-client  Cannot start service vpn-client: Address already in use
```

The Dockerfile for the strongSwan image we are using exposes ports [1].
It's unclear why this conflicts, but my theory is that we are running
charon at start which will bind to those ports, and if that happens
before the client starts, the docker daemon will not start the
vpn-client container.

This moves the vpn-client container to start before the strongswan
container to resolve this issue.

Fixes #252

[1] https://github.com/strongX509/docker/blob/master/strongswan/Dockerfile#L36

Signed-off-by: Kyle Mestery <mestery@mestery.com>